quick fix for duplicated loading
