+++ src/frontend/staff-portal/src/app/shared/components/header/header.component.ts
@@ -1,14 +1,10 @@
-import { Component, ChangeDetectionStrategy, Output, EventEmitter, OnInit } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { NavigationEnd, Router, Scroll } from '@angular/router';
-import { LoggerService } from '@core/services/logger.service';
-import { Store } from '@ngrx/store';
 import { TranslateService } from '@ngx-translate/core';
 import { AppConfigService } from 'app/services/app-config.service';
 import { AuthService, KeycloakProfile } from 'app/services/auth.service';
-import { LookupsService } from 'app/services/lookups.service';
-import { AppState, JJDisputeStore } from 'app/store';
-import { filter, forkJoin } from 'rxjs';
+import { filter } from 'rxjs';
 
 @Component({
   selector: 'app-header',
@@ -28,7 +24,6 @@ export class HeaderComponent implements OnInit {
   environment: string;
   version: string;
   isLoggedIn: Boolean = false;
-  isInit = true;
   userProfile: KeycloakProfile = {};
   jjRole: boolean = false;
   vtcRole: boolean = false;
@@ -40,9 +35,6 @@ export class HeaderComponent implements OnInit {
     private router: Router,
     private title: Title,
     private authService: AuthService,
-    private lookupsService: LookupsService,
-    private logger: LoggerService,
-    private store: Store<AppState>,
   ) {
     this.languageCode = this.translateService.getDefaultLang();
     this.onLanguage();
@@ -62,27 +54,6 @@ export class HeaderComponent implements OnInit {
       this.authService.isLoggedIn$.subscribe(isLoggedIn => {
         this.isLoggedIn = isLoggedIn;
         this.fullName = this.authService.userProfile?.fullName;
-
-        if (this.isLoggedIn) {
-          this.authService.userProfile$.subscribe(() => {
-            if (this.isInit) {
-              this.isInit = false;
-              let observables = [
-                this.authService.loadUsersLists(),
-                this.lookupsService.init()
-              ];
-
-              forkJoin(observables).subscribe({
-                next: results => {
-                  this.store.dispatch(JJDisputeStore.Actions.Get());
-                },
-                error: err => {
-                  this.logger.error("Landing Page Init: Initial data loading failed");
-                }
-              });
-            }
-          })
-        }
       })
     })
   }
